Use `compiler` syntax and drop runtime libraries
Makes use of the new `compiler` syntax with `conda-build` 3. Simplifies
the recipe a bit by dropping kludgy hacks to make sure the compilers
will work. Also drops the compiler runtime libraries as they are added
in automatically with proper version constraints using `run_exports`.
Leaves the `run` section in, but leaves it empty.