Disable building INDEX64_EXT_API for better compatibility
Disable building INDEX64_EXT_API, as it includes ILP64 symbols within
the LP64 library.  We cannot do that unless all the alternative
providers do that as well, as otherwise packages can start linking to
these symbols.  We already hit this in the wild while trying to test
BLIS with LAPACK 3.12.1.  See:
https://github.com/conda-forge/blas-feedstock/pull/152#issuecomment-3613845148

Signed-off-by: Michał Górny <mgorny@quansight.com>